Move StringUtils functions outside of deprecated header
