+++ cocos/base/ccUTF8.cpp
@@ -32,6 +32,27 @@ NS_CC_BEGIN
 
 namespace StringUtils {
 
+std::string format(const char* format, ...)
+{
+#define CC_MAX_STRING_LENGTH (1024*100)
+    
+    std::string ret;
+    
+    va_list ap;
+    va_start(ap, format);
+    
+    char* buf = (char*)malloc(CC_MAX_STRING_LENGTH);
+    if (buf != nullptr)
+    {
+        vsnprintf(buf, CC_MAX_STRING_LENGTH, format, ap);
+        ret = buf;
+        free(buf);
+    }
+    va_end(ap);
+    
+    return ret;
+}
+
 /*
  * @str:    the string to search through.
  * @c:        the character to not look for.
