+++ cocos/deprecated/CCString.h
@@ -34,9 +34,12 @@ THE SOFTWARE.
 #include <stdarg.h>
 #include <string>
 #include <functional>
-#include <sstream>
 #include "deprecated/CCArray.h"
 #include "base/CCRef.h"
+
+// We need to include `StringUtils::format()` and `StringUtils::toString()`
+// for keeping the backward compatibility
+#include "base/ccUTF8.h"
 
 NS_CC_BEGIN
 
@@ -206,20 +209,6 @@ struct StringCompare : public std::binary_function<__String *, __String *, bool>
 
 #define StringMake(str) String::create(str)
 #define ccs             StringMake
-
-namespace StringUtils {
-
-template<typename T>
-std::string toString(T arg)
-{
-    std::stringstream ss;
-    ss << arg;
-    return ss.str();
-}
-
-std::string CC_DLL format(const char* format, ...) CC_FORMAT_PRINTF(1, 2);
-    
-} // namespace StringUtils {
 
 // end of data_structure group
 /// @}
